+++ biomass/analysis/__init__.py
@@ -1,4 +1,3 @@
-from .signaling_metric import get_signaling_metric, dlnyi_dlnxj
-from .reaction import ReactionSensitivity
 from .initial_condition import InitialConditionSensitivity
 from .parameter import ParameterSensitivity
+from .reaction import ReactionSensitivity
