--- biomass/analysis/initial_condition/__init__.py
@@ -1,1 +1,0 @@
-from .sensitivity import InitialConditionSensitivity
