+++ biomass/analysis/parameter.py
@@ -1,27 +1,38 @@
 import os
 import sys
 from dataclasses import dataclass
-from typing import Dict, List
+from typing import List
 
 import matplotlib.pyplot as plt
 import numpy as np
 import seaborn as sns
 
-from ...exec_model import ExecModel, ModelObject
-from .. import dlnyi_dlnxj, get_signaling_metric
+from ..exec_model import ExecModel, ModelObject
+from .util import dlnyi_dlnxj, get_signaling_metric
 
 
 @dataclass
-class ReactionSensitivity(ExecModel):
-    """Sensitivity for rate equations"""
+class ParameterSensitivity(ExecModel):
+    """Sensitivity for parameters"""
 
     model: ModelObject
+
+    def _get_param_indices(self, excluded_params: List[str]) -> List[int]:
+        param_indices = []
+        x = self.model.pval()
+        for i, val in enumerate(x):
+            if self.model.parameters[i] not in excluded_params and val != 0.0:
+                param_indices.append(i)
+        if not param_indices:
+            raise ValueError("No nonzero parameters.")
+
+        return param_indices
 
     def _calc_sensitivity_coefficients(
         self,
         metric: str,
-        reaction_indices: List[int],
-        options: dict,
+        param_indices: List[int],
+        options,
     ) -> np.ndarray:
         """Calculating Sensitivity Coefficients
 
@@ -30,20 +41,22 @@ class ReactionSensitivity(ExecModel):
         metric : str
             The signaling metric used for sensitivity analysis.
 
-        reaction_indices : list of int
-            List of reaction indices.
+        param_indices : list of int
+            List of parameter indices for sensitivity analysis.
 
         Returns
         -------
         sensitivity_coefficients : numpy array
 
         """
+
         rate = 1.01  # 1% change
         n_file = self.get_executable()
+
         signaling_metric = np.full(
             (
                 len(n_file),
-                len(reaction_indices) + 1,
+                len(param_indices) + 1,
                 len(self.model.obs),
                 len(self.model.sim.conditions),
             ),
@@ -51,15 +64,10 @@ class ReactionSensitivity(ExecModel):
         )
         for i, nth_paramset in enumerate(n_file):
             optimized = self.load_param(nth_paramset)
-            for j, rxn_idx in enumerate(reaction_indices):
-                perturbation: Dict[int, float] = {}
-                for idx in reaction_indices:
-                    perturbation[idx] = 1.0
-                perturbation[rxn_idx] = rate
-                if (
-                    self.model.sim.simulate(optimized.params, optimized.initials, perturbation)
-                    is None
-                ):
+            for j, idx in enumerate(param_indices):
+                x = optimized.params[:]
+                x[idx] = optimized.params[idx] * rate
+                if self.model.sim.simulate(x, optimized.initials) is None:
                     for k, _ in enumerate(self.model.obs):
                         for l, _ in enumerate(self.model.sim.conditions):
                             signaling_metric[i, j, k, l] = get_signaling_metric(
@@ -67,11 +75,12 @@ class ReactionSensitivity(ExecModel):
                             )
                 sys.stdout.write(
                     "\r{:d} / {:d}".format(
-                        i * len(reaction_indices) + j + 1,
-                        len(n_file) * len(reaction_indices),
+                        i * len(param_indices) + j + 1, len(n_file) * len(param_indices)
                     )
                 )
-            if self.model.sim.simulate(optimized.params, optimized.initials) is None:
+            # Signaling metric without perturbation (j=-1)
+            x = optimized.params[:]
+            if self.model.sim.simulate(x, optimized.initials) is None:
                 for k, _ in enumerate(self.model.obs):
                     for l, _ in enumerate(self.model.sim.conditions):
                         signaling_metric[i, -1, k, l] = get_signaling_metric(
@@ -80,7 +89,7 @@ class ReactionSensitivity(ExecModel):
         sensitivity_coefficients = dlnyi_dlnxj(
             signaling_metric,
             n_file,
-            reaction_indices,
+            param_indices,
             self.model.obs,
             self.model.sim.conditions,
             rate,
@@ -91,7 +100,7 @@ class ReactionSensitivity(ExecModel):
     def _load_sc(
         self,
         metric: str,
-        reaction_indices: List[int],
+        param_indices: List[int],
         options: dict,
     ) -> np.ndarray:
         """
@@ -101,7 +110,7 @@ class ReactionSensitivity(ExecModel):
             os.path.join(
                 self.model.path,
                 "sensitivity_coefficients",
-                "reaction",
+                "parameter",
                 f"{metric}",
                 "sc.npy",
             )
@@ -110,21 +119,21 @@ class ReactionSensitivity(ExecModel):
                 os.path.join(
                     self.model.path,
                     "sensitivity_coefficients",
-                    "reaction",
+                    "parameter",
                     f"{metric}",
                 ),
                 exist_ok=True,
             )
             sensitivity_coefficients = self._calc_sensitivity_coefficients(
                 metric,
-                reaction_indices,
+                param_indices,
                 options,
             )
             np.save(
                 os.path.join(
                     self.model.path,
                     "sensitivity_coefficients",
-                    "reaction",
+                    "parameter",
                     f"{metric}",
                     "sc",
                 ),
@@ -135,78 +144,37 @@ class ReactionSensitivity(ExecModel):
                 os.path.join(
                     self.model.path,
                     "sensitivity_coefficients",
-                    "reaction",
+                    "parameter",
                     f"{metric}",
                     "sc.npy",
                 )
             )
+            if len(param_indices) != sensitivity_coefficients.shape[1]:
+                # User changed excluded_params after the last trial
+                sensitivity_coefficients = self._calc_sensitivity_coefficients(
+                    metric,
+                    param_indices,
+                    options,
+                )
+                np.save(
+                    os.path.join(
+                        self.model.path,
+                        "sensitivity_coefficients",
+                        "parameter",
+                        f"{metric}",
+                        "sc",
+                    ),
+                    sensitivity_coefficients,
+                )
 
         return sensitivity_coefficients
-
-    @staticmethod
-    def _draw_vertical_span(
-        biological_processes: List[List[int]],
-        width: float,
-    ) -> None:
-        """
-        Draw vertical span separating biological processes.
-        """
-        if len(biological_processes) > 1:
-            left_end = 0
-            for i, proc in enumerate(biological_processes):
-                if i % 2 == 0:
-                    plt.axvspan(
-                        left_end - width,
-                        left_end - width + len(proc),
-                        facecolor="k",
-                        alpha=0.1,
-                    )
-                left_end += len(proc)
-
-    def _write_reaction_indices(
-        self,
-        reaction_indices: List[int],
-        average: np.ndarray,
-        stdev: np.ndarray,
-        width: float,
-    ) -> None:
-        """
-        Put reaction index on each bar.
-        """
-        distance = np.max(average) * 0.05
-        for i, j in enumerate(reaction_indices):
-            xp = i + width * 0.5 * (len(self.model.sim.conditions) - 1)
-            yp = average[i, np.argmax(np.abs(average[i, :]))]
-            yerr = stdev[i, np.argmax(stdev[i, :])]
-            if yp > 0:
-                plt.text(
-                    xp,
-                    yp + yerr + distance,
-                    str(j),
-                    ha="center",
-                    va="bottom",
-                    fontsize=10,
-                    rotation=90,
-                )
-            else:
-                plt.text(
-                    xp,
-                    yp - yerr - distance,
-                    str(j),
-                    ha="center",
-                    va="top",
-                    fontsize=10,
-                    rotation=90,
-                )
 
     def _barplot_sensitivity(
         self,
         metric: str,
         sensitivity_coefficients: np.ndarray,
-        biological_processes: List[List[int]],
-        reaction_indices: List[int],
+        param_indices: List[int],
         save_format: str,
-        show_indices: bool,
     ) -> None:
         """
         Visualize sensitivity coefficients using barplot.
@@ -216,7 +184,7 @@ class ReactionSensitivity(ExecModel):
                 self.model.path,
                 "figure",
                 "sensitivity",
-                "reaction",
+                "parameter",
                 f"{metric}",
                 "barplot",
             ),
@@ -234,27 +202,24 @@ class ReactionSensitivity(ExecModel):
             )
         for k, obs_name in enumerate(self.model.obs):
             plt.figure(figsize=options["figsize"])
-            self._draw_vertical_span(biological_processes, options["width"])
-
-            sensitivity_array = sensitivity_coefficients[:, :, k, :]
-            # Remove NaN
-            nan_idx = []
-            for i in range(sensitivity_array.shape[0]):
-                for j in range(sensitivity_array.shape[1]):
-                    if np.isnan(sensitivity_array[i, j, :]).any():
-                        nan_idx.append(i)
-            sensitivity_array = np.delete(sensitivity_array, nan_idx, axis=0)
-            if sensitivity_array.size != 0:
-                average = np.mean(sensitivity_array, axis=0)
-                if sensitivity_array.shape[0] == 1:
-                    stdev = np.zeros((sensitivity_array.shape[1], sensitivity_array.shape[2]))
-                else:
-                    stdev = np.std(sensitivity_array, axis=0, ddof=1)
-                for l, condition in enumerate(self.model.sim.conditions):
+            plt.hlines([0], -options["width"], len(param_indices), "k", lw=1)
+            for l, condition in enumerate(self.model.sim.conditions):
+                sensitivity_matrix = sensitivity_coefficients[:, :, k, l]
+                nan_idx = []
+                for m in range(sensitivity_matrix.shape[0]):
+                    if np.isnan(sensitivity_matrix[m, :]).any():
+                        nan_idx.append(m)
+                sensitivity_matrix = np.delete(sensitivity_matrix, nan_idx, axis=0)
+                if sensitivity_matrix.size != 0:
+                    average = np.mean(sensitivity_matrix, axis=0)
+                    if sensitivity_matrix.shape[0] == 1:
+                        stdev = np.zeros(sensitivity_matrix.shape[1])
+                    else:
+                        stdev = np.std(sensitivity_matrix, axis=0, ddof=1)
                     plt.bar(
-                        np.arange(len(reaction_indices)) + l * options["width"],
-                        average[:, l],
-                        yerr=stdev[:, l],
+                        np.arange(len(param_indices)) + l * options["width"],
+                        average,
+                        yerr=stdev,
                         ecolor=options["cmap"][l],
                         capsize=2,
                         width=options["width"],
@@ -262,31 +227,32 @@ class ReactionSensitivity(ExecModel):
                         align="center",
                         label=condition,
                     )
-                if show_indices:
-                    self._write_reaction_indices(
-                        reaction_indices, average, stdev, options["width"]
-                    )
-                plt.hlines([0], -options["width"], len(reaction_indices), "k", lw=1)
-                plt.xticks([])
-                plt.ylabel(
-                    "Control coefficients on\n" + metric + " (" + obs_name.replace("_", " ") + ")"
-                )
-                plt.xlim(-options["width"], len(reaction_indices))
-                plt.legend(loc=options["legend_loc"], frameon=False)
-                plt.savefig(
-                    os.path.join(
-                        self.model.path,
-                        "figure",
-                        "sensitivity",
-                        "reaction",
-                        f"{metric}",
-                        "barplot",
-                        f"{obs_name}." + save_format,
-                    ),
-                    dpi=600 if save_format == "png" else None,
-                    bbox_inches="tight",
-                )
-                plt.close()
+            plt.xticks(
+                np.arange(len(param_indices))
+                + options["width"] * 0.5 * (len(self.model.sim.conditions) - 1),
+                [self.model.parameters[i] for i in param_indices],
+                fontsize=6,
+                rotation=90,
+            )
+            plt.ylabel(
+                "Control coefficients on\n" + metric + " (" + obs_name.replace("_", " ") + ")"
+            )
+            plt.xlim(-options["width"], len(param_indices))
+            plt.legend(loc=options["legend_loc"], frameon=False)
+            plt.savefig(
+                os.path.join(
+                    self.model.path,
+                    "figure",
+                    "sensitivity",
+                    "parameter",
+                    f"{metric}",
+                    "barplot",
+                    f"{obs_name}." + save_format,
+                ),
+                dpi=600 if save_format == "png" else None,
+                bbox_inches="tight",
+            )
+            plt.close()
 
     @staticmethod
     def _remove_nan(sensitivity_matrix: np.ndarray, normalize: bool) -> np.ndarray:
@@ -312,7 +278,7 @@ class ReactionSensitivity(ExecModel):
         self,
         metric: str,
         sensitivity_coefficients: np.ndarray,
-        reaction_indices: List[int],
+        param_indices: List[int],
         save_format: str,
     ) -> None:
         """
@@ -323,7 +289,7 @@ class ReactionSensitivity(ExecModel):
                 self.model.path,
                 "figure",
                 "sensitivity",
-                "reaction",
+                "parameter",
                 f"{metric}",
                 "heatmap",
             ),
@@ -339,7 +305,7 @@ class ReactionSensitivity(ExecModel):
                     sensitivity_coefficients[:, :, k, l], normalize=False
                 )
                 if sensitivity_matrix.shape[0] > 1 and not np.all(sensitivity_matrix == 0.0):
-                    sns.clustermap(
+                    g = sns.clustermap(
                         data=sensitivity_matrix,
                         center=0,
                         robust=True,
@@ -348,16 +314,17 @@ class ReactionSensitivity(ExecModel):
                         linewidth=0.5,
                         col_cluster=False,
                         figsize=options["figsize"],
-                        xticklabels=[str(j) for j in reaction_indices],
+                        xticklabels=[self.model.parameters[i] for i in param_indices],
                         yticklabels=[],
                         # cbar_kws={"ticks": [-1, 0, 1]}
                     )
+                    plt.setp(g.ax_heatmap.get_xticklabels(), rotation=90)
                     plt.savefig(
                         os.path.join(
                             self.model.path,
                             "figure",
                             "sensitivity",
-                            "reaction",
+                            "parameter",
                             f"{metric}",
                             "heatmap",
                             f"{condition}_{obs_name}." + save_format,
@@ -371,51 +338,24 @@ class ReactionSensitivity(ExecModel):
         """
         Perform sensitivity analysis.
         """
-        if not self.model.rxn.reactions:
-            raise ValueError("Define reaction indices (reactions) in reaction_network.py")
-        biological_processes = self.model.rxn.group()
-        reaction_indices = np.sum(biological_processes, axis=0)
-        sensitivity_coefficients = self._load_sc(metric, reaction_indices, options)
-
+        param_indices = self._get_param_indices(options["excluded_params"])
+        sensitivity_coefficients = self._load_sc(metric, param_indices, options)
         if style == "barplot":
             self._barplot_sensitivity(
                 metric,
                 sensitivity_coefficients,
-                biological_processes,
-                reaction_indices,
+                param_indices,
                 save_format,
-                options["show_indices"],
             )
         elif style == "heatmap":
-            self._heatmap_sensitivity(
-                metric,
-                sensitivity_coefficients,
-                reaction_indices,
-                save_format,
-            )
+            if len(param_indices) < 2:
+                pass
+            else:
+                self._heatmap_sensitivity(
+                    metric,
+                    sensitivity_coefficients,
+                    param_indices,
+                    save_format,
+                )
         else:
             raise ValueError("Available styles are: 'barplot', 'heatmap'")
-
-
-def is_duplicate(
-    reactions: Dict[str, List[int]],
-    biological_processes: List[List[int]],
-) -> bool:
-    reaction_indices = (
-        sum(biological_processes, []) if len(reactions) > 1 else biological_processes[0]
-    )
-    duplicate_reaction = [i for i in set(reaction_indices) if reaction_indices.count(i) > 1]
-    if not duplicate_reaction:
-        return False
-    else:
-        which_process = []
-        for reaction_index in duplicate_reaction:
-            for process, indices in reactions.items():
-                if reaction_index in indices:
-                    which_process.append(process)
-        raise ValueError(
-            "Duplicate reaction: {} found in {}.".format(
-                ", ".join(map(str, duplicate_reaction)),
-                ", ".join(which_process),
-            )
-        )
