--- biomass/analysis/parameter/__init__.py
@@ -1,1 +1,0 @@
-from .sensitivity import ParameterSensitivity
