+++ biomass/analysis/reaction.py
@@ -1,38 +1,27 @@
 import os
 import sys
 from dataclasses import dataclass
-from typing import List
+from typing import Dict, List
 
 import matplotlib.pyplot as plt
 import numpy as np
 import seaborn as sns
 
-from ...exec_model import ExecModel, ModelObject
-from .. import dlnyi_dlnxj, get_signaling_metric
+from ..exec_model import ExecModel, ModelObject
+from .util import dlnyi_dlnxj, get_signaling_metric
 
 
 @dataclass
-class ParameterSensitivity(ExecModel):
-    """Sensitivity for parameters"""
+class ReactionSensitivity(ExecModel):
+    """Sensitivity for rate equations"""
 
     model: ModelObject
-
-    def _get_param_indices(self, excluded_params: List[str]) -> List[int]:
-        param_indices = []
-        x = self.model.pval()
-        for i, val in enumerate(x):
-            if self.model.parameters[i] not in excluded_params and val != 0.0:
-                param_indices.append(i)
-        if not param_indices:
-            raise ValueError("No nonzero parameters.")
-
-        return param_indices
 
     def _calc_sensitivity_coefficients(
         self,
         metric: str,
-        param_indices: List[int],
-        options,
+        reaction_indices: List[int],
+        options: dict,
     ) -> np.ndarray:
         """Calculating Sensitivity Coefficients
 
@@ -41,22 +30,20 @@ class ParameterSensitivity(ExecModel):
         metric : str
             The signaling metric used for sensitivity analysis.
 
-        param_indices : list of int
-            List of parameter indices for sensitivity analysis.
+        reaction_indices : list of int
+            List of reaction indices.
 
         Returns
         -------
         sensitivity_coefficients : numpy array
 
         """
-
         rate = 1.01  # 1% change
         n_file = self.get_executable()
-
         signaling_metric = np.full(
             (
                 len(n_file),
-                len(param_indices) + 1,
+                len(reaction_indices) + 1,
                 len(self.model.obs),
                 len(self.model.sim.conditions),
             ),
@@ -64,10 +51,15 @@ class ParameterSensitivity(ExecModel):
         )
         for i, nth_paramset in enumerate(n_file):
             optimized = self.load_param(nth_paramset)
-            for j, idx in enumerate(param_indices):
-                x = optimized.params[:]
-                x[idx] = optimized.params[idx] * rate
-                if self.model.sim.simulate(x, optimized.initials) is None:
+            for j, rxn_idx in enumerate(reaction_indices):
+                perturbation: Dict[int, float] = {}
+                for idx in reaction_indices:
+                    perturbation[idx] = 1.0
+                perturbation[rxn_idx] = rate
+                if (
+                    self.model.sim.simulate(optimized.params, optimized.initials, perturbation)
+                    is None
+                ):
                     for k, _ in enumerate(self.model.obs):
                         for l, _ in enumerate(self.model.sim.conditions):
                             signaling_metric[i, j, k, l] = get_signaling_metric(
@@ -75,12 +67,11 @@ class ParameterSensitivity(ExecModel):
                             )
                 sys.stdout.write(
                     "\r{:d} / {:d}".format(
-                        i * len(param_indices) + j + 1, len(n_file) * len(param_indices)
+                        i * len(reaction_indices) + j + 1,
+                        len(n_file) * len(reaction_indices),
                     )
                 )
-            # Signaling metric without perturbation (j=-1)
-            x = optimized.params[:]
-            if self.model.sim.simulate(x, optimized.initials) is None:
+            if self.model.sim.simulate(optimized.params, optimized.initials) is None:
                 for k, _ in enumerate(self.model.obs):
                     for l, _ in enumerate(self.model.sim.conditions):
                         signaling_metric[i, -1, k, l] = get_signaling_metric(
@@ -89,7 +80,7 @@ class ParameterSensitivity(ExecModel):
         sensitivity_coefficients = dlnyi_dlnxj(
             signaling_metric,
             n_file,
-            param_indices,
+            reaction_indices,
             self.model.obs,
             self.model.sim.conditions,
             rate,
@@ -100,7 +91,7 @@ class ParameterSensitivity(ExecModel):
     def _load_sc(
         self,
         metric: str,
-        param_indices: List[int],
+        reaction_indices: List[int],
         options: dict,
     ) -> np.ndarray:
         """
@@ -110,7 +101,7 @@ class ParameterSensitivity(ExecModel):
             os.path.join(
                 self.model.path,
                 "sensitivity_coefficients",
-                "parameter",
+                "reaction",
                 f"{metric}",
                 "sc.npy",
             )
@@ -119,21 +110,21 @@ class ParameterSensitivity(ExecModel):
                 os.path.join(
                     self.model.path,
                     "sensitivity_coefficients",
-                    "parameter",
+                    "reaction",
                     f"{metric}",
                 ),
                 exist_ok=True,
             )
             sensitivity_coefficients = self._calc_sensitivity_coefficients(
                 metric,
-                param_indices,
+                reaction_indices,
                 options,
             )
             np.save(
                 os.path.join(
                     self.model.path,
                     "sensitivity_coefficients",
-                    "parameter",
+                    "reaction",
                     f"{metric}",
                     "sc",
                 ),
@@ -144,37 +135,78 @@ class ParameterSensitivity(ExecModel):
                 os.path.join(
                     self.model.path,
                     "sensitivity_coefficients",
-                    "parameter",
+                    "reaction",
                     f"{metric}",
                     "sc.npy",
                 )
             )
-            if len(param_indices) != sensitivity_coefficients.shape[1]:
-                # User changed excluded_params after the last trial
-                sensitivity_coefficients = self._calc_sensitivity_coefficients(
-                    metric,
-                    param_indices,
-                    options,
-                )
-                np.save(
-                    os.path.join(
-                        self.model.path,
-                        "sensitivity_coefficients",
-                        "parameter",
-                        f"{metric}",
-                        "sc",
-                    ),
-                    sensitivity_coefficients,
-                )
 
         return sensitivity_coefficients
+
+    @staticmethod
+    def _draw_vertical_span(
+        biological_processes: List[List[int]],
+        width: float,
+    ) -> None:
+        """
+        Draw vertical span separating biological processes.
+        """
+        if len(biological_processes) > 1:
+            left_end = 0
+            for i, proc in enumerate(biological_processes):
+                if i % 2 == 0:
+                    plt.axvspan(
+                        left_end - width,
+                        left_end - width + len(proc),
+                        facecolor="k",
+                        alpha=0.1,
+                    )
+                left_end += len(proc)
+
+    def _write_reaction_indices(
+        self,
+        reaction_indices: List[int],
+        average: np.ndarray,
+        stdev: np.ndarray,
+        width: float,
+    ) -> None:
+        """
+        Put reaction index on each bar.
+        """
+        distance = np.max(average) * 0.05
+        for i, j in enumerate(reaction_indices):
+            xp = i + width * 0.5 * (len(self.model.sim.conditions) - 1)
+            yp = average[i, np.argmax(np.abs(average[i, :]))]
+            yerr = stdev[i, np.argmax(stdev[i, :])]
+            if yp > 0:
+                plt.text(
+                    xp,
+                    yp + yerr + distance,
+                    str(j),
+                    ha="center",
+                    va="bottom",
+                    fontsize=10,
+                    rotation=90,
+                )
+            else:
+                plt.text(
+                    xp,
+                    yp - yerr - distance,
+                    str(j),
+                    ha="center",
+                    va="top",
+                    fontsize=10,
+                    rotation=90,
+                )
 
     def _barplot_sensitivity(
         self,
         metric: str,
         sensitivity_coefficients: np.ndarray,
-        param_indices: List[int],
+        biological_processes: List[List[int]],
+        reaction_indices: List[int],
         save_format: str,
+        show_indices: bool,
     ) -> None:
         """
         Visualize sensitivity coefficients using barplot.
@@ -184,7 +216,7 @@ class ParameterSensitivity(ExecModel):
                 self.model.path,
                 "figure",
                 "sensitivity",
-                "parameter",
+                "reaction",
                 f"{metric}",
                 "barplot",
             ),
@@ -202,24 +234,27 @@ class ParameterSensitivity(ExecModel):
             )
         for k, obs_name in enumerate(self.model.obs):
             plt.figure(figsize=options["figsize"])
-            plt.hlines([0], -options["width"], len(param_indices), "k", lw=1)
-            for l, condition in enumerate(self.model.sim.conditions):
-                sensitivity_matrix = sensitivity_coefficients[:, :, k, l]
-                nan_idx = []
-                for m in range(sensitivity_matrix.shape[0]):
-                    if np.isnan(sensitivity_matrix[m, :]).any():
-                        nan_idx.append(m)
-                sensitivity_matrix = np.delete(sensitivity_matrix, nan_idx, axis=0)
-                if sensitivity_matrix.size != 0:
-                    average = np.mean(sensitivity_matrix, axis=0)
-                    if sensitivity_matrix.shape[0] == 1:
-                        stdev = np.zeros(sensitivity_matrix.shape[1])
-                    else:
-                        stdev = np.std(sensitivity_matrix, axis=0, ddof=1)
+            self._draw_vertical_span(biological_processes, options["width"])
+
+            sensitivity_array = sensitivity_coefficients[:, :, k, :]
+            # Remove NaN
+            nan_idx = []
+            for i in range(sensitivity_array.shape[0]):
+                for j in range(sensitivity_array.shape[1]):
+                    if np.isnan(sensitivity_array[i, j, :]).any():
+                        nan_idx.append(i)
+            sensitivity_array = np.delete(sensitivity_array, nan_idx, axis=0)
+            if sensitivity_array.size != 0:
+                average = np.mean(sensitivity_array, axis=0)
+                if sensitivity_array.shape[0] == 1:
+                    stdev = np.zeros((sensitivity_array.shape[1], sensitivity_array.shape[2]))
+                else:
+                    stdev = np.std(sensitivity_array, axis=0, ddof=1)
+                for l, condition in enumerate(self.model.sim.conditions):
                     plt.bar(
-                        np.arange(len(param_indices)) + l * options["width"],
-                        average,
-                        yerr=stdev,
+                        np.arange(len(reaction_indices)) + l * options["width"],
+                        average[:, l],
+                        yerr=stdev[:, l],
                         ecolor=options["cmap"][l],
                         capsize=2,
                         width=options["width"],
@@ -227,32 +262,31 @@ class ParameterSensitivity(ExecModel):
                         align="center",
                         label=condition,
                     )
-            plt.xticks(
-                np.arange(len(param_indices))
-                + options["width"] * 0.5 * (len(self.model.sim.conditions) - 1),
-                [self.model.parameters[i] for i in param_indices],
-                fontsize=6,
-                rotation=90,
-            )
-            plt.ylabel(
-                "Control coefficients on\n" + metric + " (" + obs_name.replace("_", " ") + ")"
-            )
-            plt.xlim(-options["width"], len(param_indices))
-            plt.legend(loc=options["legend_loc"], frameon=False)
-            plt.savefig(
-                os.path.join(
-                    self.model.path,
-                    "figure",
-                    "sensitivity",
-                    "parameter",
-                    f"{metric}",
-                    "barplot",
-                    f"{obs_name}." + save_format,
-                ),
-                dpi=600 if save_format == "png" else None,
-                bbox_inches="tight",
-            )
-            plt.close()
+                if show_indices:
+                    self._write_reaction_indices(
+                        reaction_indices, average, stdev, options["width"]
+                    )
+                plt.hlines([0], -options["width"], len(reaction_indices), "k", lw=1)
+                plt.xticks([])
+                plt.ylabel(
+                    "Control coefficients on\n" + metric + " (" + obs_name.replace("_", " ") + ")"
+                )
+                plt.xlim(-options["width"], len(reaction_indices))
+                plt.legend(loc=options["legend_loc"], frameon=False)
+                plt.savefig(
+                    os.path.join(
+                        self.model.path,
+                        "figure",
+                        "sensitivity",
+                        "reaction",
+                        f"{metric}",
+                        "barplot",
+                        f"{obs_name}." + save_format,
+                    ),
+                    dpi=600 if save_format == "png" else None,
+                    bbox_inches="tight",
+                )
+                plt.close()
 
     @staticmethod
     def _remove_nan(sensitivity_matrix: np.ndarray, normalize: bool) -> np.ndarray:
@@ -278,7 +312,7 @@ class ParameterSensitivity(ExecModel):
         self,
         metric: str,
         sensitivity_coefficients: np.ndarray,
-        param_indices: List[int],
+        reaction_indices: List[int],
         save_format: str,
     ) -> None:
         """
@@ -289,7 +323,7 @@ class ParameterSensitivity(ExecModel):
                 self.model.path,
                 "figure",
                 "sensitivity",
-                "parameter",
+                "reaction",
                 f"{metric}",
                 "heatmap",
             ),
@@ -305,7 +339,7 @@ class ParameterSensitivity(ExecModel):
                     sensitivity_coefficients[:, :, k, l], normalize=False
                 )
                 if sensitivity_matrix.shape[0] > 1 and not np.all(sensitivity_matrix == 0.0):
-                    g = sns.clustermap(
+                    sns.clustermap(
                         data=sensitivity_matrix,
                         center=0,
                         robust=True,
@@ -314,17 +348,16 @@ class ParameterSensitivity(ExecModel):
                         linewidth=0.5,
                         col_cluster=False,
                         figsize=options["figsize"],
-                        xticklabels=[self.model.parameters[i] for i in param_indices],
+                        xticklabels=[str(j) for j in reaction_indices],
                         yticklabels=[],
                         # cbar_kws={"ticks": [-1, 0, 1]}
                     )
-                    plt.setp(g.ax_heatmap.get_xticklabels(), rotation=90)
                     plt.savefig(
                         os.path.join(
                             self.model.path,
                             "figure",
                             "sensitivity",
-                            "parameter",
+                            "reaction",
                             f"{metric}",
                             "heatmap",
                             f"{condition}_{obs_name}." + save_format,
@@ -338,24 +371,51 @@ class ParameterSensitivity(ExecModel):
         """
         Perform sensitivity analysis.
         """
-        param_indices = self._get_param_indices(options["excluded_params"])
-        sensitivity_coefficients = self._load_sc(metric, param_indices, options)
+        if not self.model.rxn.reactions:
+            raise ValueError("Define reaction indices (reactions) in reaction_network.py")
+        biological_processes = self.model.rxn.group()
+        reaction_indices = np.sum(biological_processes, axis=0)
+        sensitivity_coefficients = self._load_sc(metric, reaction_indices, options)
+
         if style == "barplot":
             self._barplot_sensitivity(
                 metric,
                 sensitivity_coefficients,
-                param_indices,
+                biological_processes,
+                reaction_indices,
                 save_format,
+                options["show_indices"],
             )
         elif style == "heatmap":
-            if len(param_indices) < 2:
-                pass
-            else:
-                self._heatmap_sensitivity(
-                    metric,
-                    sensitivity_coefficients,
-                    param_indices,
-                    save_format,
-                )
+            self._heatmap_sensitivity(
+                metric,
+                sensitivity_coefficients,
+                reaction_indices,
+                save_format,
+            )
         else:
             raise ValueError("Available styles are: 'barplot', 'heatmap'")
+
+
+def is_duplicate(
+    reactions: Dict[str, List[int]],
+    biological_processes: List[List[int]],
+) -> bool:
+    reaction_indices = (
+        sum(biological_processes, []) if len(reactions) > 1 else biological_processes[0]
+    )
+    duplicate_reaction = [i for i in set(reaction_indices) if reaction_indices.count(i) > 1]
+    if not duplicate_reaction:
+        return False
+    else:
+        which_process = []
+        for reaction_index in duplicate_reaction:
+            for process, indices in reactions.items():
+                if reaction_index in indices:
+                    which_process.append(process)
+        raise ValueError(
+            "Duplicate reaction: {} found in {}.".format(
+                ", ".join(map(str, duplicate_reaction)),
+                ", ".join(which_process),
+            )
+        )
