--- biomass/analysis/reaction/__init__.py
@@ -1,1 +1,0 @@
-from .sensitivity import ReactionSensitivity, is_duplicate
